Add MoviesController with index action
This commit introduces a new "Movies" controller with an index action. This action populates an @movies instance variable with a hard-coded array of movie titles. Changes in routes have been made to direct "/movies" and root URLs to the movies#index action. Additionally, a simple view is also created to list all movies.