Add show action for movies
A show route has been added to the routes configuration, focused on extracting the ID from the URL and passing it to the new show action within the movies controller. This action then finds a movie by the provided ID and renders a new view, which displays detailed information about a specific movie.