Add navigation links to application
A name (as: :movie) has been added to the 'movies#show' route in the 'config/routes.rb' file for improved readability. Additionally, navigation links have been incorporated into the views - linking movie titles on the index page to their show pages, and the added navigation bar in the application layout file includes a link to the 'All Movies' page and a link to the root_path on the logo image.