Add movie editing functionality
A new route, controller action, and view file have been added to enable movie editing. The edit action was added to the movies controller and a corresponding route added to the routes file. An edit view was also created with a form for editing a movie's details. Additionally, an "Edit" button was added to the movie details page.